Move mock to __tests__
